refactor: 폴더 이름 변경 utils -> functions
